Move generic_objects tests under the automate directory
