+++ cfme/tests/automate/generic_objects/buttons/test_generic_objects.py
