+++ cfme/tests/automate/generic_objects/test_instances.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 import fauxfactory
 import pytest
-import yaml
 
 import cfme.rest.gen_data as rest_gen_data
 from cfme import test_requirements
-from cfme.base.login import BaseLoggedInPage
 from cfme.generic_objects.instance.ui import MyServiceGenericObjectInstanceView
 from cfme.services.myservice import MyService
 from cfme.tests.automate.custom_button import TextInputDialogView
@@ -18,22 +16,6 @@ from cfme.utils.wait import wait_for
 
 
 pytestmark = [test_requirements.generic_objects]
-
-
-GEN_OBJ_DIRECTORY = "/var/www/miq/vmdb/tmp/generic_object_definitions"
-
-
-@pytest.fixture
-def gen_obj_def_import_export(appliance):
-    with appliance.context.use(ViaREST):
-        definition = appliance.collections.generic_object_definitions.create(
-            name="rest_gen_class_imp_exp{}".format(fauxfactory.gen_alphanumeric()),
-            description="Generic Object Definition",
-            attributes={'addr01': 'string'},
-            methods=['add_vm', 'remove_vm']
-        )
-        yield definition
-        definition.delete_if_exists()
 
 
 @pytest.fixture(scope="module")
@@ -61,45 +43,6 @@ def button_with_dialog(appliance, generic_object, dialog):
         yield button
 
         button.delete_if_exists()
-
-
-@pytest.fixture(scope="module")
-def generic_object_button_group(appliance, generic_definition):
-    def _generic_object_button_group(create_action=True):
-        if create_action:
-            with appliance.context.use(ViaUI):
-                group_name = "button_group_{}".format(fauxfactory.gen_alphanumeric())
-                group_desc = "Group_button_description_{}".format(fauxfactory.gen_alphanumeric())
-                groups_buttons = generic_definition.collections.generic_object_groups_buttons
-                generic_object_button_group = groups_buttons.create(
-                    name=group_name, description=group_desc, image="fa-user"
-                )
-                view = appliance.browser.create_view(BaseLoggedInPage)
-                view.flash.assert_no_error()
-            return generic_object_button_group
-
-    return _generic_object_button_group
-
-
-@pytest.fixture(scope="module")
-def generic_object_button(appliance, generic_object_button_group, generic_definition):
-    def _generic_object_button(button_group):
-        with appliance.context.use(ViaUI):
-            button_parent = (
-                generic_object_button_group(button_group) if button_group else generic_definition
-            )
-            button_name = 'button_{}'.format(fauxfactory.gen_alphanumeric())
-            button_desc = 'Button_description_{}'.format(fauxfactory.gen_alphanumeric())
-            generic_object_button = button_parent.collections.generic_object_buttons.create(
-                name=button_name,
-                description=button_desc,
-                image='fa-home',
-                request=fauxfactory.gen_alphanumeric()
-            )
-            view = appliance.browser.create_view(BaseLoggedInPage)
-            view.flash.assert_no_error()
-        return generic_object_button
-    return _generic_object_button
 
 
 @pytest.mark.sauce
@@ -161,34 +104,6 @@ def test_generic_objects_crud(appliance, context, request):
         assert not instance.exists
 
 
-@pytest.mark.parametrize('button_group', [True, False],
-                         ids=['button_group_with_button', 'single_button'])
-def test_generic_objects_with_buttons_ui(appliance, request, add_generic_object_to_service,
-                                         button_group, generic_object_button):
-    """
-        Tests buttons ui visibility assigned to generic object
-
-        Metadata:
-            test_flag: ui
-
-    Polarion:
-        assignee: jdupuy
-        initialEstimate: 1/4h
-        casecomponent: GenericObjects
-    """
-    instance = add_generic_object_to_service
-    generic_button = generic_object_button(button_group)
-    generic_button_group = generic_button.parent.parent
-
-    with appliance.context.use(ViaUI):
-        view = navigate_to(instance, 'MyServiceDetails')
-        if button_group:
-            assert view.toolbar.group(generic_button_group.name).custom_button.has_item(
-                generic_button.name)
-        else:
-            assert view.toolbar.button(generic_button.name).custom_button.is_displayed
-
-
 @test_requirements.tag
 @pytest.mark.parametrize('tag_place', [True, False], ids=['details', 'collection'])
 def test_generic_objects_tag_ui(appliance, generic_object, tag_place):
@@ -237,74 +152,6 @@ def test_generic_objects_tag_rest(appliance, generic_object, tags):
         generic_object.remove_tag(tag)
         tag_available = generic_object.get_tags()
         assert tag.id not in [t.id for t in tag_available]
-
-
-@pytest.mark.ignore_stream("5.10")
-def test_import_export_generic_object_definition(request, appliance, gen_obj_def_import_export):
-    """
-    Bugzilla:
-        1595259
-
-    Polarion:
-        assignee: jdupuy
-        initialEstimate: 1/6h
-        caseimportance: high
-        caseposneg: positive
-        testtype: functional
-        startsin: 5.11
-        casecomponent: GenericObjects
-        testSteps:
-            1. Create generic object definition via Rest
-            2. Export the generic object definition
-            3. Delete the generic object definition
-            4. Import the generic object definition
-        expectedResults:
-            1. The generic object definition should be present in CFME
-            2. Yaml file should be present on the appliance with the generic object details
-            3. Generic object definition is deleted
-            4. Generic object definition once again exists on the appliance
-    """
-    # Create the generic object directory
-    assert appliance.ssh_client.run_command("mkdir {}".format(GEN_OBJ_DIRECTORY)).success
-
-    @request.addfinalizer
-    def cleanup():
-        assert appliance.ssh_client.run_command("rm -rf {}".format(GEN_OBJ_DIRECTORY)).success
-
-    # Export the user defined generic object definitions
-    assert appliance.ssh_client.run_rake_command(
-        "evm:export:generic_object_definitions -- --directory {}".format(GEN_OBJ_DIRECTORY)
-    ).success
-    # Verify the file's information
-    try:
-        with appliance.ssh_client.open_sftp().open(
-                "{}/{}.yaml".format(GEN_OBJ_DIRECTORY, gen_obj_def_import_export.name)
-        ) as f:
-            data = yaml.safe_load(f)[0]["GenericObjectDefinition"]
-
-    except IOError:
-        pytest.fail(
-            "IOError: {}/{}.yaml not found on the appliance, "
-            "exporting the generic object definition failed".format(
-                GEN_OBJ_DIRECTORY, gen_obj_def_import_export.name
-            )
-        )
-
-    assert data.get("description") == gen_obj_def_import_export.description
-    assert data.get("name") == gen_obj_def_import_export.name
-
-    # Delete the generic object definition via the UI
-    gen_obj_def_import_export.delete_if_exists()
-
-    # Import the generic object yaml by running the rake command
-    assert appliance.ssh_client.run_rake_command(
-        "evm:import:generic_object_definitions -- --source {}/{}.yaml".format(
-            GEN_OBJ_DIRECTORY,
-            gen_obj_def_import_export.name,
-        )
-    ).success
-
-    assert gen_obj_def_import_export.exists
 
 
 @test_requirements.customer_stories
